Renamed directory helper to common
