+++ coollime/src/main/java/coollime/binarytree/bst/getkeys/Solution.java
@@ -1,6 +1,6 @@
 package coollime.binarytree.bst.getkeys;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.ArrayList;
 import java.util.Deque;
 import java.util.LinkedList;
