+++ coollime/src/main/java/coollime/binarytree/bst/search/Solution.java
@@ -1,6 +1,6 @@
 package coollime.binarytree.bst.search;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 
 public class Solution {
