+++ coollime/src/main/java/coollime/binarytree/bst/validbst/Solution.java
@@ -1,6 +1,6 @@
 package coollime.binarytree.bst.validbst;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 
 public class Solution {
