+++ coollime/src/main/java/coollime/binarytree/general/symmetric/Solution.java
@@ -1,6 +1,6 @@
 package coollime.binarytree.general.symmetric;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 
 public class Solution {
