+++ coollime/src/main/java/coollime/binarytree/traversal/inorder/Solution.java
@@ -1,6 +1,6 @@
 package coollime.binarytree.traversal.inorder;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Deque;
