+++ coollime/src/main/java/coollime/binarytree/traversal/postorder/Solution.java
@@ -1,6 +1,6 @@
 package coollime.binarytree.traversal.postorder;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.ArrayList;
 import java.util.Deque;
 import java.util.LinkedList;
