+++ coollime/src/main/java/coollime/binarytree/traversal/preorder/Solution.java
@@ -1,6 +1,6 @@
 package coollime.binarytree.traversal.preorder;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.ArrayList;
 import java.util.Deque;
 import java.util.LinkedList;
