+++ coollime/src/main/java/coollime/common/objects/GraphNode.java
@@ -1,10 +1,11 @@
-package coollime.helper;
+package coollime.common.objects;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 public class GraphNode {
+
   public int key;
   public List<GraphNode> neighbors;
 
@@ -14,22 +15,24 @@ public class GraphNode {
   }
 
   /**
-   * Build an undirected graph given a list of keys and adjacency list of each node
+   * Build an undirected graph given a list of keys and adjacency list of each
+   * node
    *
-   * Structure of adjacency list:
-   * [ [neighbors of node 0], [neighbors of node 1], [neighbors of node 2], ... , ]
+   * Structure of adjacency list: [ [neighbors of node 0], [neighbors of node 1],
+   * [neighbors of node 2], ... , ]
    *
    * The keys list should have the same length as that of adjacency list
    *
    * @param nodeKeys      Keys of corresponding nodes
    * @param adjacencyList Adjacency list of each node
-   * @return              An undirected graph of the nodes
+   * @return An undirected graph of the nodes
    */
   public static List<GraphNode> buildGraph(
     List<Integer> nodeKeys,
     List<List<GraphNode>> adjacencyList
   ) {
-    // TODO(MZ): This is not correct - Need to figure out another way (probably need a map or set)
+    // TODO(MZ): This is not correct - Need to figure out another way (probably need
+    // a map or set)
     List<GraphNode> graph = new ArrayList<>();
     if (nodeKeys == null || nodeKeys.isEmpty()) {
       // return graph;
