+++ coollime/src/main/java/coollime/linkedlist/checkiflinkedlistispalindrome/Solution.java
@@ -1,14 +1,14 @@
 package coollime.linkedlist.checkiflinkedlistispalindrome;
 
-import coollime.helper.ListNode;
+import coollime.common.objects.ListNode;
 
 public class Solution {
 
   /**
    * Check whether a linked list is a palindrome
    *
-   * @param head  The head node of the input linked list
-   * @return      Whether the linked list is a palindrome
+   * @param head The head node of the input linked list
+   * @return Whether the linked list is a palindrome
    */
   public boolean isPalindrome(ListNode head) {
     if (head == null || head.next == null) {
@@ -35,10 +35,10 @@ public class Solution {
   /**
    * Find the middle node of a linked list
    *
-   * @param head  The head node of the linked list
-   * @return      The middle node of the linked list
-   *              In the case where there are two middle nodes in the list (when
-   *              there are even number of nodes), return the first one
+   * @param head The head node of the linked list
+   * @return The middle node of the linked list In the case where there are two
+   *         middle nodes in the list (when there are even number of nodes),
+   *         return the first one
    */
   private ListNode findMiddleNode(ListNode head) {
     if (head == null || head.next == null) {
@@ -56,8 +56,9 @@ public class Solution {
   /**
    * Reverse a singly-linked list
    *
-   * @param head  The head node of the linked list
-   * @return      The new linked list that is a reversed version of the input linked list
+   * @param head The head node of the linked list
+   * @return The new linked list that is a reversed version of the input linked
+   *         list
    */
   private ListNode reverse(ListNode head) {
     if (head == null) {
