+++ coollime/src/main/java/coollime/linkedlist/reorderlinkedlist/Solution.java
@@ -1,6 +1,6 @@
 package coollime.linkedlist.reorderlinkedlist;
 
-import coollime.helper.ListNode;
+import coollime.common.objects.ListNode;
 
 public class Solution {
