+++ coollime/src/main/java/coollime/recursion/maximumpathsumbinarytreeiii/Solution.java
@@ -1,6 +1,6 @@
 package coollime.recursion.maximumpathsumbinarytreeiii;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 
 public class Solution {
