+++ coollime/src/test/java/coollime/binarytree/bst/insert/InsertInBinarySearchTreeTest.java
@@ -2,7 +2,7 @@ package coollime.binarytree.bst.insert;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.Arrays;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
