+++ coollime/src/test/java/coollime/binarytree/bst/search/SearchInBinarySearchTreeTest.java
@@ -3,7 +3,7 @@ package coollime.binarytree.bst.search;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.Arrays;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
