+++ coollime/src/test/java/coollime/binarytree/bst/validbst/IsBinarySearchTreeOrNotTest.java
@@ -3,7 +3,7 @@ package coollime.binarytree.bst.validbst;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.Arrays;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
