+++ coollime/src/test/java/coollime/binarytree/general/tweaked/TweakedIdenticalBinaryTreesTest.java
@@ -3,7 +3,7 @@ package coollime.binarytree.general.tweaked;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.Arrays;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
