+++ coollime/src/test/java/coollime/binarytree/path/sumi/BinaryTreePathSumToTargetITest.java
@@ -3,7 +3,7 @@ package coollime.binarytree.path.sumi;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.Arrays;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
