+++ coollime/src/test/java/coollime/binarytree/traversal/preorder/PreorderTraversalOfBinaryTreeTest.java
@@ -2,7 +2,7 @@ package coollime.binarytree.traversal.preorder;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
