+++ coollime/src/test/java/coollime/linkedlist/insertinsortedlinkedlist/InsertInSortedLinkedListTest.java
@@ -2,7 +2,7 @@ package coollime.linkedlist.insertinsortedlinkedlist;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import coollime.helper.ListNode;
+import coollime.common.objects.ListNode;
 import java.util.ArrayList;
 import java.util.Arrays;
 import org.junit.jupiter.api.AfterEach;
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class InsertInSortedLinkedListTest {
+
   private static Solution sut;
   private ListNode head;
   private int value;
