+++ coollime/src/test/java/coollime/linkedlist/middlenodeoflinkedlist/MiddleNodeOfLinkedListTest.java
@@ -2,7 +2,7 @@ package coollime.linkedlist.middlenodeoflinkedlist;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import coollime.helper.ListNode;
+import coollime.common.objects.ListNode;
 import java.util.ArrayList;
 import java.util.Arrays;
 import org.junit.jupiter.api.AfterEach;
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class MiddleNodeOfLinkedListTest {
+
   private static Solution sut;
   private ListNode head;
