+++ coollime/src/test/java/coollime/recursion/binarytreepathsumtotargetiii/BinaryTreePathSumToTargetIIITest.java
@@ -2,7 +2,7 @@ package coollime.recursion.binarytreepathsumtotargetiii;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.ArrayList;
 import java.util.Arrays;
 import org.junit.jupiter.api.BeforeAll;
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class BinaryTreePathSumToTargetIIITest {
+
   private static Solution sut;
   private TreeNode root;
   private int target;
