+++ coollime/src/test/java/coollime/recursion/maximumpathsumbinarytreeiii/MaximumPathSumBinaryTreeIIITest.java
@@ -2,7 +2,7 @@ package coollime.recursion.maximumpathsumbinarytreeiii;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import coollime.helper.TreeNode;
+import coollime.common.objects.TreeNode;
 import java.util.ArrayList;
 import java.util.Arrays;
 import org.junit.jupiter.api.BeforeAll;
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class MaximumPathSumBinaryTreeIIITest {
+
   private static Solution sut;
   private TreeNode root;
